refactor(agent-task): extract lab-offload recording from lifecycle_ops (#9927)
