+++ crates/homeboy-agents/src/agent_task_lifecycle/mod.rs
@@ -39,6 +39,7 @@ mod cook_workspace_restore;
 mod failure_recording;
 mod health;
 mod lab_handoff_reconciliation;
+mod lab_offload;
 mod lifecycle_candidate_adoption;
 mod lifecycle_ops;
 mod lifecycle_record_ops;
@@ -56,6 +57,7 @@ pub use failure_recording::*;
 pub use health::*;
 pub use homeboy_core::controller_runtime::ControllerRuntimePruneResult;
 pub use lab_handoff_reconciliation::*;
+pub use lab_offload::*;
 pub use lifecycle_candidate_adoption::*;
 pub use lifecycle_ops::*;
 pub use lifecycle_record_ops::cook_attempt_run_id;
